py_lift/pyproject.toml
[tool.poetry]
name = "py_lift"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11, <3.12"
dkpro-cassis = "^0.10.1"
pytest = "^7.4.3"
pyspellchecker = "^0.8.1"
textstat = "^0.7.3"
pytest-cov = "^4.1.0"
wordfreq = "^3.1.1"
pandas = "^2.2.0"
polars = "1.33.1"
udapi = "^0.3.0"
pyconll = "^3.2.0"
pyarrow = "^15.0.0"
setuptools = "80.4.0"
cas-visualizer = "^0.1.1"
jupyter = "^1.1.1"
Deprecated = "^1.2.18"
lift-resources-lists = {path = "../shared_resources/resources/lift_resources_lists", develop = true}
streamlit = "^1.50.0"
spacy = "^3.8.7"
de-core-news-sm = { url = "https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.8.0/de_core_news_sm-3.8.0-py3-none-any.whl" }
en-core-web-md = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "."
]
